+++ HISTORY.md
@@ -4,6 +4,7 @@
 30th January 2025 - 0.26.0-beta2
 --------------------------------
 
+ * added support (including coloured output) for (in)equality testing of UDTs;
  * now ensures that does not emit ANSI colour sequences when running on Windows on an old terminal that does not support them;